+++ Conv/Section/Section+Delegate.swift
@@ -1,0 +1,21 @@
+//
+//  Section+Delegate.swift
+//  Conv
+//
+//  Created by Yudai.Hirose on 2018/07/25.
+//  Copyright © 2018年 廣瀬雄大. All rights reserved.
+//
+
+import Foundation
+
+extension Section: SectionDelegatable {
+    public func inset(collectionView: UICollectionView, collectionViewLayout: UICollectionViewLayout, section: Int) -> UIEdgeInsets? {
+        return inset?((self, collectionView, collectionViewLayout, section))
+    }
+    public func minimumLineSpacing(collectionView: UICollectionView, collectionViewLayout: UICollectionViewLayout, section: Int) -> CGFloat? {
+        return minimumLineSpacing?((self, collectionView, collectionViewLayout, section))
+    }
+    public func minimumInteritemSpacing(collectionView: UICollectionView, collectionViewLayout: UICollectionViewLayout, section: Int) -> CGFloat? {
+        return minimumInteritemSpacing?((self, collectionView, collectionViewLayout, section))
+    }
+}
