+++ Conv/Section/SectionDelegate.swift
@@ -1,0 +1,16 @@
+//
+//  SectionDelegate.swift
+//  Conv
+//
+//  Created by Yudai.Hirose on 2018/07/25.
+//  Copyright © 2018年 廣瀬雄大. All rights reserved.
+//
+
+import Foundation
+
+public protocol SectionDelegatable {
+    func inset(collectionView: UICollectionView, collectionViewLayout: UICollectionViewLayout, section: Int) -> UIEdgeInsets?
+    func minimumLineSpacing(collectionView: UICollectionView, collectionViewLayout: UICollectionViewLayout, section: Int) -> CGFloat?
+    func minimumInteritemSpacing(collectionView: UICollectionView, collectionViewLayout: UICollectionViewLayout, section: Int) -> CGFloat?
+}
+
